modularization + update of constants some fixes
