+++ src/WvsGame/Maple/Commands/Implementation/JobCommand.cs
@@ -1,5 +1,6 @@
 ﻿using Destiny.Maple.Characters;
 using System;
+using Destiny.Constants;
 
 namespace Destiny.Maple.Commands.Implementation
 {
@@ -41,9 +42,9 @@ namespace Destiny.Maple.Commands.Implementation
                 {
                     short jobID = short.Parse(args[0]);
 
-                    if (Enum.IsDefined(typeof(Job), jobID))
+                    if (Enum.IsDefined(typeof(CharacterConstants.Job), jobID))
                     {
-                        caller.Job = (Job)jobID;
+                        caller.Job = (CharacterConstants.Job)jobID;
                     }
                     else
                     {
@@ -54,7 +55,7 @@ namespace Destiny.Maple.Commands.Implementation
                 {
                     try
                     {
-                        caller.Job = (Job)Enum.Parse(typeof(Job), args[0], true);
+                        caller.Job = (CharacterConstants.Job)Enum.Parse(typeof(CharacterConstants.Job), args[0], true);
                     }
                     catch (ArgumentException)
                     {
