+++ src/WvsGame/Maple/Experience.cs
@@ -1,5 +1,6 @@
 ﻿using Destiny.Maple.Characters;
 using Destiny.Network;
+using Destiny.Constants;
 
 namespace Destiny.Maple
 {
@@ -17,7 +18,7 @@ namespace Destiny.Maple
 
         public static int GetExpNeededForLevel(int level)
         {
-            return level > 200 ? 2000000000 : ExperienceTables.CharacterLevel[level];
+            return level > 200 ? 2000000000 : CharacterConstants.ExperienceTables.CharacterLevel[level];
         }
 
         public static void giveExp(Character character, int exp)
