+++ src/WvsLogin/Maple/Account.cs
@@ -3,6 +3,7 @@ using Destiny.IO;
 using Destiny.Network;
 using System;
 using System.Data;
+using Destiny.Constants;
 
 namespace Destiny.Maple
 {
@@ -15,7 +16,7 @@ namespace Destiny.Maple
         public string Password { get; set; }
         public string Salt { get; set; }
         public bool EULA { get; set; }
-        public Gender Gender { get; set; }
+        public CharacterConstants.Gender Gender { get; set; }
         public string Pin { get; set; }
         public string Pic { get; set; }
         public bool IsBanned { get; set; }
@@ -51,7 +52,7 @@ namespace Destiny.Maple
             this.Password = (string)datum["Password"];
             this.Salt = (string)datum["Salt"];
             this.EULA = (bool)datum["EULA"];
-            this.Gender = (Gender)datum["Gender"];
+            this.Gender = (CharacterConstants.Gender)datum["Gender"];
             this.Pin = (string)datum["Pin"];
             this.Pic = (string)datum["Pic"];
             this.IsBanned = (bool)datum["IsBanned"];
