+++ src/WvsLogin/Network/LoginClient.cs
@@ -4,6 +4,7 @@ using Destiny.Security;
 using System;
 using System.Collections.Generic;
 using System.Net.Sockets;
+using Destiny.Constants;
 
 namespace Destiny.Network
 {
@@ -147,7 +148,7 @@ namespace Destiny.Network
                         this.Account.Salt = HashGenerator.GenerateMD5();
                         this.Account.Password = SHACryptograph.Encrypt(SHAMode.SHA512, password + this.Account.Salt);
                         this.Account.EULA = false;
-                        this.Account.Gender = Gender.Unset;
+                        this.Account.Gender = CharacterConstants.Gender.Unset;
                         this.Account.Pin = string.Empty;
                         this.Account.Pic = string.Empty;
                         this.Account.IsBanned = false;
@@ -226,7 +227,7 @@ namespace Destiny.Network
 
         private void SetGender(Packet iPacket)
         {
-            if (this.Account.Gender != Gender.Unset)
+            if (this.Account.Gender != CharacterConstants.Gender.Unset)
             {
                 return;
             }
@@ -235,7 +236,7 @@ namespace Destiny.Network
 
             if (valid)
             {
-                Gender gender = (Gender)iPacket.ReadByte();
+                CharacterConstants.Gender gender = (CharacterConstants.Gender)iPacket.ReadByte();
 
                 this.Account.Gender = gender;
